backend/requirements.txt
fastapi
uvicorn
pytest
requests
python-dotenv
